+++ Sources/TuistKit/GraphMappers/GraphMapperProvider.swift
@@ -3,6 +3,7 @@ import TuistCache
 import TuistCloud
 import TuistCore
 import TuistSigning
+import TuistGenerator
 
 /// It defines an interface for providing the mappers to be used for a specific configuration.
 protocol GraphMapperProviding {