+++ Tests/TuistKitTests/GraphMappers/GraphMapperProviderTests.swift
@@ -4,6 +4,7 @@ import TuistSigning
 import TuistCloud
 import TuistCore
 import TuistCoreTesting
+import TuistGenerator
 import TuistSupport
 import XCTest